move generated application entities into own subpackage
