+++ bundles/org.openhab.binding.saicismart/pom.xml
@@ -120,11 +120,11 @@
                 <argument>-m</argument>
                 <argument>java</argument>
                 <argument>-o</argument>
-                <argument>${project.basedir}/target/generated-sources/asn1/org/openhab/binding/saicismart/internal/asn1/v1_1</argument>
+                <argument>${project.basedir}/target/generated-sources/asn1/org/openhab/binding/saicismart/internal/asn1/v1_1/entity</argument>
                 <argument>-f</argument>
                 <argument>${project.basedir}/src/main/asn1/v1_1/ApplicationData.asn1</argument>
                 <argument>-ns</argument>
-                <argument>org.openhab.binding.saicismart.internal.asn1.v1_1</argument>
+                <argument>org.openhab.binding.saicismart.internal.asn1.v1_1.entity</argument>
               </arguments>
             </configuration>
           </execution>
@@ -180,11 +180,11 @@
                 <argument>-m</argument>
                 <argument>java</argument>
                 <argument>-o</argument>
-                <argument>${project.basedir}/target/generated-sources/asn1/org/openhab/binding/saicismart/internal/asn1/v2_1</argument>
+                <argument>${project.basedir}/target/generated-sources/asn1/org/openhab/binding/saicismart/internal/asn1/v2_1/entity</argument>
                 <argument>-f</argument>
                 <argument>${project.basedir}/src/main/asn1/v2_1/ApplicationData.asn1</argument>
                 <argument>-ns</argument>
-                <argument>org.openhab.binding.saicismart.internal.asn1.v2_1</argument>
+                <argument>org.openhab.binding.saicismart.internal.asn1.v2_1.entity</argument>
               </arguments>
             </configuration>
           </execution>
@@ -240,11 +240,11 @@
                 <argument>-m</argument>
                 <argument>java</argument>
                 <argument>-o</argument>
-                <argument>${project.basedir}/target/generated-sources/asn1/org/openhab/binding/saicismart/internal/asn1/v3_0</argument>
+                <argument>${project.basedir}/target/generated-sources/asn1/org/openhab/binding/saicismart/internal/asn1/v3_0/entity</argument>
                 <argument>-f</argument>
                 <argument>${project.basedir}/src/main/asn1/v3_0/ApplicationData.asn1</argument>
                 <argument>-ns</argument>
-                <argument>org.openhab.binding.saicismart.internal.asn1.v3_0</argument>
+                <argument>org.openhab.binding.saicismart.internal.asn1.v3_0.entity</argument>
               </arguments>
             </configuration>
           </execution>
