+++ bundles/org.openhab.binding.saicismart/src/main/java/org/openhab/binding/saicismart/internal/SAICiSMARTBridgeHandler.java
@@ -27,12 +27,12 @@ import org.eclipse.jetty.client.HttpClient;
 import org.eclipse.jetty.client.util.StringContentProvider;
 import org.openhab.binding.saicismart.internal.asn1.v1_1.MP_DispatcherBody;
 import org.openhab.binding.saicismart.internal.asn1.v1_1.MP_DispatcherHeader;
-import org.openhab.binding.saicismart.internal.asn1.v1_1.MP_UserLoggingInReq;
-import org.openhab.binding.saicismart.internal.asn1.v1_1.MP_UserLoggingInResp;
 import org.openhab.binding.saicismart.internal.asn1.v1_1.Message;
 import org.openhab.binding.saicismart.internal.asn1.v1_1.MessageCoder;
 import org.openhab.binding.saicismart.internal.asn1.v1_1.MessageCounter;
-import org.openhab.binding.saicismart.internal.asn1.v1_1.VinInfo;
+import org.openhab.binding.saicismart.internal.asn1.v1_1.entity.MP_UserLoggingInReq;
+import org.openhab.binding.saicismart.internal.asn1.v1_1.entity.MP_UserLoggingInResp;
+import org.openhab.binding.saicismart.internal.asn1.v1_1.entity.VinInfo;
 import org.openhab.core.io.net.http.HttpClientFactory;
 import org.openhab.core.thing.Bridge;
 import org.openhab.core.thing.ChannelUID;
