+++ bundles/org.openhab.binding.saicismart/src/main/java/org/openhab/binding/saicismart/internal/VehicleStateUpdater.java
@@ -34,8 +34,8 @@ import org.openhab.binding.saicismart.internal.asn1.v2_1.MP_DispatcherBody;
 import org.openhab.binding.saicismart.internal.asn1.v2_1.MP_DispatcherHeader;
 import org.openhab.binding.saicismart.internal.asn1.v2_1.Message;
 import org.openhab.binding.saicismart.internal.asn1.v2_1.MessageCoder;
-import org.openhab.binding.saicismart.internal.asn1.v2_1.OTA_RVMVehicleStatusReq;
-import org.openhab.binding.saicismart.internal.asn1.v2_1.OTA_RVMVehicleStatusResp25857;
+import org.openhab.binding.saicismart.internal.asn1.v2_1.entity.OTA_RVMVehicleStatusReq;
+import org.openhab.binding.saicismart.internal.asn1.v2_1.entity.OTA_RVMVehicleStatusResp25857;
 import org.openhab.core.library.types.DecimalType;
 import org.openhab.core.library.types.OnOffType;
 import org.openhab.core.library.types.OpenClosedType;
